# openapi.yaml
openapi: 3.1.0
info:
  title: GanoScan Backend
  description: Ganoderma detection inference + scan history for the GanoScan Android app.
  version: 1.0.0
servers:
- url: http://127.0.0.1:5005
  description: Local development
paths:
  /health:
    get:
      tags:
      - meta
      summary: Service + model status
      description: Always reachable without an API key, for load balancer / uptime checks.
      operationId: getHealth
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /predict:
    post:
      tags:
      - inference
      summary: Classify an uploaded photo and persist the scan
      operationId: predict
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_predict'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResult'
        '400':
          description: Empty or unreadable image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '413':
          description: Image exceeds the configured size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Missing/invalid form fields (FastAPI validation)
        '500':
          description: Model inference failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      security:
      - APIKeyHeader: []
  /history:
    get:
      tags:
      - scans
      summary: List scans, newest first
      operationId: listHistory
      security:
      - APIKeyHeader: []
      parameters:
      - name: verdict
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(HEALTHY|INFECTED)$
          - type: 'null'
          title: Verdict
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /scan/{scan_id}:
    get:
      tags:
      - scans
      summary: Get one scan by id
      operationId: getScan
      security:
      - APIKeyHeader: []
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResult'
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No scan with this id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - scans
      summary: Delete one scan and its stored photo
      operationId: deleteScan
      security:
      - APIKeyHeader: []
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      - name: x-admin-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteOneResult'
        '401':
          description: Missing/invalid X-API-Key, or (if configured) X-Admin-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No scan with this id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /stats:
    get:
      tags:
      - scans
      summary: Aggregate scan counts
      operationId: getStats
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      security:
      - APIKeyHeader: []
  /scans:
    delete:
      tags:
      - scans
      summary: Bulk delete scans (and their photos)
      operationId: deleteScans
      security:
      - APIKeyHeader: []
      parameters:
      - name: verdict
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(HEALTHY|INFECTED)$
          - type: 'null'
          title: Verdict
      - name: before
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ISO datetime; deletes scans created before this
          title: Before
        description: ISO datetime; deletes scans created before this
      - name: all
        in: query
        required: false
        schema:
          type: boolean
          description: Required (=true) to delete with no filters at all
          default: false
          title: All
        description: Required (=true) to delete with no filters at all
      - name: x-admin-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteManyResult'
        '400':
          description: No filters given and all=true not passed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '401':
          description: Missing/invalid X-API-Key, or (if configured) X-Admin-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_predict:
      properties:
        image:
          type: string
          format: binary
          title: Image
          description: Leaf/stem photo (jpg, jpeg, png, bmp or webp)
        treeId:
          anyOf:
          - type: string
          - type: 'null'
          title: Treeid
          description: 'e.g. ''Pohon #A-142''; defaults to ''Pohon #<generated id>'''
        block:
          anyOf:
          - type: string
          - type: 'null'
          title: Block
          description: e.g. 'Blok C'; defaults to GANOSCAN_DEFAULT_BLOCK
      type: object
      required:
      - image
      title: Body_predict
    DeleteManyResult:
      properties:
        deleted:
          type: integer
          title: Deleted
      type: object
      required:
      - deleted
      title: DeleteManyResult
    DeleteOneResult:
      properties:
        deleted:
          type: boolean
          title: Deleted
        id:
          type: string
          title: Id
      type: object
      required:
      - deleted
      - id
      title: DeleteOneResult
    ErrorDetail:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorDetail
      description: Shape of every error response (FastAPI's default HTTPException body).
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Health:
      properties:
        status:
          type: string
          title: Status
        service:
          type: string
          title: Service
        version:
          type: string
          title: Version
        model:
          $ref: '#/components/schemas/ModelInfo'
      type: object
      required:
      - status
      - service
      - version
      - model
      title: Health
    HistoryResponse:
      properties:
        scans:
          items:
            $ref: '#/components/schemas/ScanResult'
          type: array
          title: Scans
        count:
          type: integer
          title: Count
      type: object
      required:
      - scans
      - count
      title: HistoryResponse
    ModelInfo:
      properties:
        mode:
          type: string
          title: Mode
        loaded:
          type: boolean
          title: Loaded
        backbone:
          type: string
          title: Backbone
        classes:
          items:
            type: string
          type: array
          title: Classes
        imageSize:
          items:
            type: integer
          type: array
          title: Imagesize
        device:
          type: string
          title: Device
        testAccuracy:
          anyOf:
          - type: number
          - type: 'null'
          title: Testaccuracy
      type: object
      required:
      - mode
      - loaded
      - backbone
      - classes
      - imageSize
      - device
      title: ModelInfo
    ScanResult:
      properties:
        id:
          type: string
          title: Id
        treeId:
          type: string
          title: Treeid
        block:
          type: string
          title: Block
        verdict:
          type: string
          title: Verdict
        label:
          type: string
          title: Label
        predictedClass:
          type: string
          title: Predictedclass
        severity:
          type: string
          title: Severity
        confidence:
          type: number
          title: Confidence
        probabilities:
          additionalProperties:
            type: number
          type: object
          title: Probabilities
        gamma:
          type: number
          title: Gamma
        inputResolution:
          type: string
          title: Inputresolution
        recommendations:
          items:
            type: string
          type: array
          title: Recommendations
        stages:
          items:
            $ref: '#/components/schemas/Stage'
          type: array
          title: Stages
        imageUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Imageurl
        createdAt:
          type: string
          title: Createdat
        time:
          type: string
          title: Time
        dateLabel:
          type: string
          title: Datelabel
        dayGroup:
          type: string
          title: Daygroup
        mock:
          type: boolean
          title: Mock
      type: object
      required:
      - id
      - treeId
      - block
      - verdict
      - label
      - predictedClass
      - severity
      - confidence
      - probabilities
      - gamma
      - inputResolution
      - recommendations
      - stages
      - createdAt
      - time
      - dateLabel
      - dayGroup
      - mock
      title: ScanResult
    Stage:
      properties:
        index:
          type: string
          title: Index
        title:
          type: string
          title: Title
        sub:
          type: string
          title: Sub
        done:
          type: boolean
          title: Done
          default: false
      type: object
      required:
      - index
      - title
      - sub
      title: Stage
    Stats:
      properties:
        totalScan:
          type: integer
          title: Totalscan
        totalHealthy:
          type: integer
          title: Totalhealthy
        totalInfected:
          type: integer
          title: Totalinfected
      type: object
      required:
      - totalScan
      - totalHealthy
      - totalInfected
      title: Stats
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key